refactor: #169 Refactor Board Types
modules/board/types 파일의 타입들이 많아짐에 따라
분류 작업의 필요성을 느꼈습니다.

Resolves #169
