+++ application/millage-client/src/modules/board/types/entities/Comment.ts
@@ -1,0 +1,5 @@
+/* Types of Comment */
+export type Comment = {
+    id: number;
+    content: string;
+};
