+++ application/millage-client/src/modules/board/types/entities/Poll.ts
@@ -1,0 +1,10 @@
+/* Types of Poll */
+export type Poll = {
+    id: number;
+    content: string;
+    votes: number;
+};
+
+export type PollInputs = Pick<Poll, 'content'> & {
+    index: number;
+};
