+++ application/millage-client/src/modules/board/types/entities/Recruit.ts
@@ -1,0 +1,6 @@
+/* Types of Recruit */
+export type RecruitStatus = {
+    status: 'progress' | 'completed'
+    totalMember: number;
+    currentMember: number;
+};
