+++ application/millage-client/src/modules/board/types/entities/index.ts
@@ -1,0 +1,5 @@
+export * from './Board';
+export * from './Post';
+export * from './Comment';
+export * from './Poll';
+export * from './Recruit';
